A clean separation by extracting the methods to a base class
